Partial fix to build issue
